+++ smart-contract/program/src/processor/unstake.rs
@@ -5,12 +5,12 @@ use crate::{
 };
 use bonfida_utils::{BorshSize, InstructionsAccount};
 use borsh::{BorshDeserialize, BorshSerialize};
-use solana_program::clock::Clock;
-use solana_program::sysvar::Sysvar;
-use solana_program::{account_info::{next_account_info, AccountInfo}, entrypoint::ProgramResult, msg, program_error::ProgramError, pubkey::Pubkey};
+use solana_program::{account_info::{next_account_info, AccountInfo}, entrypoint::ProgramResult, program_error::ProgramError, pubkey::Pubkey};
+use solana_program::program::invoke_signed;
+use spl_token::instruction::transfer;
 
 use crate::error::AccessError;
-use crate::state::{BondAccount, StakeAccount, StakePool, UnstakeRequest};
+use crate::state::{BondAccount, StakeAccount, StakePool, StakePoolHeader};
 
 #[derive(BorshDeserialize, BorshSerialize, BorshSize)]
 /// The required parameters for the `unstake` instruction
@@ -39,6 +39,17 @@ pub struct Accounts<'a, T> {
     #[cons(signer)]
     pub owner: &'a T,
 
+    /// The destination of the staked tokens
+    #[cons(writable)]
+    pub destination_token: &'a T,
+
+    /// The SPL token program account
+    pub spl_token_program: &'a T,
+
+    /// The stake pool vault
+    #[cons(writable)]
+    pub vault: &'a T,
+
     /// Optional bond account to be able to stake under the minimum
     pub bond_account: Option<&'a T>,
 }
@@ -55,12 +66,22 @@ impl<'a, 'b: 'a> Accounts<'a, AccountInfo<'b>> {
             stake_account: next_account_info(accounts_iter)?,
             stake_pool: next_account_info(accounts_iter)?,
             owner: next_account_info(accounts_iter)?,
+            destination_token: next_account_info(accounts_iter)?,
+            spl_token_program: next_account_info(accounts_iter)?,
+            vault: next_account_info(accounts_iter)?,
             bond_account: if has_bond_account {
                 next_account_info(accounts_iter).ok()
             } else {
                 None
             },
         };
+
+        // Check keys
+        check_account_key(
+            accounts.spl_token_program,
+            &spl_token::ID,
+            AccessError::WrongSplTokenProgramId,
+        )?;
 
         // Check ownership
         check_account_owner(
@@ -77,6 +98,16 @@ impl<'a, 'b: 'a> Accounts<'a, AccountInfo<'b>> {
             accounts.stake_pool,
             program_id,
             AccessError::WrongStakePoolAccountOwner,
+        )?;
+        check_account_owner(
+            accounts.destination_token,
+            &spl_token::ID,
+            AccessError::WrongTokenAccountOwner,
+        )?;
+        check_account_owner(
+            accounts.vault,
+            &spl_token::ID,
+            AccessError::WrongTokenAccountOwner,
         )?;
         if let Some(bond_account) = accounts.bond_account {
             check_account_owner(
@@ -104,7 +135,6 @@ pub fn process_unstake(
     let mut stake_pool = StakePool::get_checked(accounts.stake_pool, vec![Tag::StakePool])?;
     let mut stake_account = StakeAccount::from_account_info(accounts.stake_account)?;
     let mut central_state = CentralState::from_account_info(accounts.central_state_account)?;
-    let current_time = Clock::get()?.unix_timestamp;
 
     if stake_account.last_claimed_offset < stake_pool.header.current_day_idx as i64 {
         return Err(AccessError::UnclaimedRewards.into());
@@ -119,6 +149,11 @@ pub fn process_unstake(
         accounts.stake_pool,
         &stake_account.stake_pool,
         AccessError::StakePoolMismatch,
+    )?;
+    check_account_key(
+        accounts.vault,
+        &Pubkey::new(&stake_pool.header.vault),
+        AccessError::StakePoolVaultMismatch,
     )?;
 
     let mut amount_in_bonds: u64 = 0;
@@ -156,8 +191,33 @@ pub fn process_unstake(
     stake_account.withdraw(amount)?;
     stake_pool.header.withdraw(amount, central_state.last_snapshot_offset, central_state.creation_time)?;
 
-    // Add unstake request
-    stake_account.add_unstake_request(UnstakeRequest::new(amount, current_time))?;
+    // Transfer tokens
+    let signer_seeds: &[&[u8]] = &[
+        StakePoolHeader::SEED,
+        &stake_pool.header.owner.clone(),
+        &[stake_pool.header.nonce],
+    ];
+    let transfer_instruction = transfer(
+        &spl_token::ID,
+        accounts.vault.key,
+        accounts.destination_token.key,
+        accounts.stake_pool.key,
+        &[],
+        amount,
+    )?;
+
+    drop(stake_pool);
+
+    invoke_signed(
+        &transfer_instruction,
+        &[
+            accounts.spl_token_program.clone(),
+            accounts.vault.clone(),
+            accounts.destination_token.clone(),
+            accounts.stake_pool.clone(),
+        ],
+        &[signer_seeds],
+    )?;
 
     // Save states
     stake_account.save(&mut accounts.stake_account.data.borrow_mut())?;
