+++ smart-contract/program/tests/common_stake_limit.rs
@@ -1,24 +1,14 @@
-use std::os::macos::raw::stat;
 use solana_program::{pubkey::Pubkey, system_program};
 use solana_program_test::{processor, ProgramTest};
-use solana_test_framework::*;
 use solana_sdk::signer::{keypair::Keypair, Signer};
-use solana_sdk::sysvar::{clock};
-use spl_associated_token_account::{instruction::create_associated_token_account, get_associated_token_address};
-pub mod common;
+use solana_sdk::sysvar::clock;
+use solana_test_framework::*;
+use spl_associated_token_account::{get_associated_token_address, instruction::create_associated_token_account};
+
+use crate::common::test_runner::TestRunner;
 use crate::common::utils::{mint_bootstrap, sign_send_instructions};
-use crate::common::test_runner::{TestRunner};
-use access_protocol::{
-    entrypoint::process_instruction,
-    instruction::{
-        activate_stake_pool, admin_mint,
-        claim_pool_rewards, claim_rewards,
-        crank, create_central_state, create_stake_account,
-        create_stake_pool, execute_unstake, stake, unstake,
-    },
-};
-use mpl_token_metadata::pda::find_metadata_account;
-use access_protocol::instruction::create_bond;
+
+pub mod common;
 
 #[tokio::test]
 async fn common_stake_limit() {
