+++ smart-contract/program/tests/common_unstake_limit.rs
@@ -1,24 +1,8 @@
-use std::os::macos::raw::stat;
-use solana_program::{pubkey::Pubkey, system_program};
-use solana_program_test::{processor, ProgramTest};
 use solana_test_framework::*;
-use solana_sdk::signer::{keypair::Keypair, Signer};
-use solana_sdk::sysvar::{clock};
-use spl_associated_token_account::{instruction::create_associated_token_account, get_associated_token_address};
+use solana_sdk::signer::{Signer};
 pub mod common;
-use crate::common::utils::{mint_bootstrap, sign_send_instructions};
+use crate::common::utils::{sign_send_instructions};
 use crate::common::test_runner::{TestRunner};
-use access_protocol::{
-    entrypoint::process_instruction,
-    instruction::{
-        activate_stake_pool, admin_mint,
-        claim_pool_rewards, claim_rewards,
-        crank, create_central_state, create_stake_account,
-        create_stake_pool, execute_unstake, stake, unstake,
-    },
-};
-use mpl_token_metadata::pda::find_metadata_account;
-use access_protocol::instruction::create_bond;
 
 #[tokio::test]
 async fn common_unstake_limit() {
@@ -51,7 +35,7 @@ async fn common_unstake_limit() {
     assert_eq!(pool_stats.total_pool_staked, 1100);
 
     // todo investigate why this is needed and if we can get rid of it
-    tr.sleep(1).await;
+    tr.sleep(1).await.unwrap();
 
     // unstake above the pool minimum should work
     tr.unstake(&stake_pool_owner.pubkey(), &staker, 100).await.unwrap();
@@ -59,7 +43,7 @@ async fn common_unstake_limit() {
     assert_eq!(pool_stats.total_pool_staked, 1000);
 
     // todo investigate why this is needed and if we can get rid of it
-    tr.sleep(1).await;
+    tr.sleep(1).await.unwrap();
 
     // full unstake should work
     tr.unstake(&stake_pool_owner.pubkey(), &staker, 1000).await.unwrap();
